Fabric SetBlock Mixin + Cleanup
Due to Fabric not having modules listed in stack traces, our implementation of detecting non-vanilla block setting had be be overhauled for Fabric. The previous implementation is now how the detection is handled only on Forge. On Fabric, we now use the stack trace and compare it to the loaded mods and check if a modID is within the package identifier of the stack trace, which it should be. The output of the mixin is the same on both Forge and Fabric when tested with the Create mod. Changed the database connection system around so that the same connection will be used for the entire life cycle of the loaded world. Added proper logging for errors.