Relational Block Data + Forge Mixin Fixes
Converted the blockSet data to now use the relational database setup using the table managers. Added a couple of new methods to the interface to allow callbacks when insertion has completed and also a method to allow us to clear any caches we created. The block sets have caches created for the foreign keys in the database. This is for performance so we don't need to query the database to find the foreign key every time. Typically these caches are cleared on insertion completion except for those where the number of keys is not great, such as dimension or action types for blocks. These are cleared on server start/stop and also on database reconnection, just to be safe. Fixed the anvil and hoe item mixin preventing the game from booting on Forge due to differences in method names on forge/fabric due to lambdas not having persistent names. Switched QUEUED_DATA to using a concurrent hashmap to prevent rare crashes. Introduced some more robust error logging also to help diagnose which SQL statement failed rather than just printing the error. May refactor the methods for getting cached values as they are very similar.