Multithreaded database implementation
Added basic multithreading into the database querying and committing to prevent lag spikes when transactions are occurring. This uses a thread pool of currently 4 threads. Added a check to the block mixin to prevent unnecessary processing on the client. Also added a guard to the AddToDatabase method to prevent the render thread adding data to the database.