Inspection Mode + Paginated SQL Query
Added 2 commands, inspect and page. Inspect will toggle inspection mode. This means when the player clicks on a block, they will receive information about it. The page command allows switching between pages of information. The SQL queries have been reworked to only send back a small number of sorted records, offset based on the page number. This reduces the load on the server significantly and is much faster. Retrieving from the database now returns a struct with the arraylist of records inside as well as the total number of records which match the supplied count SQL query. This is used to calculate the max page number in the inspect command. Fixed a crash where the mod would not work on dedicated server due to client side code accessing. Renamed the BlockSetBlockData variables to match those in the database. Theoretically, the infrastructure work for the database should now be complete.