BlockSetCauses + Formatted Non Player Interactions
Reworked the BlockUtils functions to be more adaptive to different types of interactions in the world. A new interface has been created called BlockSetCause which allows a tag to be set for the cause (which is stored in the database) and a format to be used when displaying the action in the inspection logs. This allows much more versatility with the messages displayed. All current actions have been swapped over to this new system.