Multiblock interactions
Interactions involving multiblocks such as doors, beds and double flowers are now logged. Only player interactions are currently logged for this, so for example an explosion that destroys a door will only be marked as the cause on one half of the door. This will be reworked and fixed when the BlockSetActions are unified into the BlockSetCause system.