Automated Testing
* Create build.yml

Added an initial build action for gradle java 17 targeted at 1.20.1.

* Reworked testing action

* Remove cache + Artifacts

Removed cache as setup gradle action already handles that. Artifacts are now also uploaded, so they can be used in the next set of testing when loading a client and server.

* Removed validation

It's included in setup gradle action.

* Updated versions

* Try without gradle wrapper

* Client loading test

Added an action to try and launch the game client. The action is now only run on push to 1.20.1 branch and on pull requests.

* Try and fix forge infinite loading + caching

* Allow manual action trigger

* Timeout test

If a test-launch job does not complete within 10 minutes now it will fail, rather than endlessly hanging. This may need to be tweaked in future.

* Download Architectury API automatically

Try to grab the architectury API from the CurseForge maven.

* Debug list downloaded files

* JarInJar JDBC Driver

The JDBC driver is included in exported build files. This should allow forge and fabric to pass the tests now.

* Clean up