Mixin Rework
Reworked all mixins completely to use WrapOperation. Some of these still need tweaking and reimplementing, especially the trees and the forge side of things.
Reworked the BlockUtils class to have methods for adding to the block database, this was mainly implemented to reduce the length of lines when adding entries. This also contains a new set of methods for use with WrapOperation mixins to ensure the block change actually happened and was not cancelled.
All mixins now have a very low priority so hopefully they will be injected last in the chain.
The modded block interaction mixin has been reworked into the level mixin on the common side, instead of two seperate implementations of the same functionality.
Forge currently does not run and there is an issue with the JDBC driver.
The nether portal mixin has also been implemented.