Mixin Tweaks
Tweaked all existing mixins which had issues to make sure they work as expected. Also redid a lot of the causes to not show the database tag in the inspection message. Added a mixin for campfires and candle extinguishing/lighting.
Everything in the mod up to this point has been tested and seems functional.