Unsupported Deduplication
Removed check for the cause when getting a timestamp to insert into the database by improving the underlying issue. Blocks "breaking" from the unsupported cause are only recorded if another negatively returning call to canSurvive on that blockPosttion was not made within 5 ticks, rather than relying on rounding issues which could occur from truncating down to seconds. Also renamed the FeatureMixinHolder to a generic class for holding any future mixin variables which are needed.