New LevelSet Mixin
Overhauled the LevelSetMixin to have some extra functionality. This may change in future as there is almost certainly changes to be made for some edge cases, particularly to do with the way how we determine if a levelSet has been tracked already as this may be a little dodgy in some circumstances.
- Any call to levelSet which has not been accounted for by another mixin wrapping it is now logged to console and also to the database as if it were a modded interaction. The idea behind this is to help catch missed block interactions in development.
- This system only tracks changes to blocks, not blockstates, to reduce the amount of spam.
- Console logging is only in a development environment, or (when implemented) the config is set accordingly.
- The mixin now correctly stores the old blockstate in the database.