Level Mixin Spam Removal
Removed some spam logging caused by chained block neighbour updates in the level mixin. This was most notable with fire as some instances of fire spreading utilise block updates. The changes introduced here may be reworked in future if this causes issues. For now, a warning is issued if a nested call to setBlock is made to ease development. This is only logged if we are in the development environment or (in future) a config is set. Updated fire block mixin to fix a missing case of fire spreading.