Localised Block Names
Block names are now localised when viewed in the inspection mode. This works with any language selected and it is resolved using the in game registry. If a block name is not found in the registry, it's internal name will be used instead.
The InspectionMode class has been cleaned up a bit, particularly when creating the mutable components to be displayed.