Localisation + Cause Rework
Implemented the startings of localisation within the mod. Most things are localised. Added a few configs to allow further customization of the output of the inspect command.
Reworked the block cause system so that all causes use a base class instead of individual implementations of the BlockSetCause interface, doing this allowed us to move the getInspectMessage logic to using the language file rather than constructing the component in code.
In the event a cause may not have a source component, the BlockSetCauseWithSourceFallback should be used instead, to allow a fallback localisation to be specified in the language file.
In the level mixin, chained neighbour updates only log block information, not blockstate to prevent cases where the console would get spammed due to the same interactions, but just different block state information.
There are a few things which still need localising, the main offender is the explosion sourceDetails field; which currently is constructed serverside then saved into the database, the localisation for this needs to be moved so it is resolved at inspection time.