Mob targeting logging
Hostile mobs involved in block changes in the world now log every entity they have been tracking in their lifespan, as opposed to just the most recent target. This is also localised using the new system, rather than recording the full text into the database. Using interface injection, we've added a method to the mob class, allowing us to get the list of targets which we add to when we change target using a normal inject mixin.