Inventory Position Tracking
Added some functions to the server player via interface injection which allows the position of the currently opened inventory to be stored and accessed. This needs to be tweaked slightly as some modded inventories do not seem to work with this system. Also made some test code to log moving items about in the inventory, this will be tracked properly soon and then a database implementation will be written for this.