Block Access Logging
Added logging for when a block with a menu is accessed, i.e. a player opens a chest, or looks in a barrel. Also fixed a bug where some inventories would instantly have their position cleared in the mixin as some menu's force close the player's inventory before rendering the new one.