Manual Player Based Inventory Injections
Implemented the various mixin injections and wrappings needed for tracking players interacting with inventory slots. This needs a bit of cleaning up, but generally works well. This only handles player interactions with inventories, so not blocks like hoppers. The database side of things for item tracking has not yet been implemented.