Inventory Logging Rewrite
Rewrote the manual slot based inventory logging implemented in the last commit, as there were various issues with it, the main ones being covering all edge cases and mod compat as some mods may change the way how clicks are handled etc. The new method of tracking inventories relies on injections into the slots themselves and tracking the changes rather than the overall menu clicks. This commit should be functionally identical to the last one, but with a lot less jank and greater readability. Also removed the list of injected interfaces in architectury's file as we cast to them anyway.