Entity Item Logging
Added a basic implementation of logging item transfers from entities such as minecarts, boats, and animals such as horses. This uses practically the same system as for blocks with inventories. There is a bug currently with horses and other ridable entities that the accessed menu is not set again when opening the inventory, but instead only happens when interacting with the ridable entity. Essentially the method for detecting opening a rideable entities inventory needs to be reworked. Removed an old unused block set cause class, and moved entity events into their own class.