Combined Block Inspection
Added the basic ability to inspect from different tables using inspection mode. This currently is crudely implemented in a new class which inspection mode calls. Ideally this shouldn't have as much duplicate code as the DataManager and this will probably be refactored somewhat soon. This commit allows both inspection of inventory changes in blocks and also block changes themselves.