A few refinements
Bumped version to 1.0.1
Added block entity interactions logging. This is different from container accesses as they previously were under the same cause.
Added a system where only 1 retrieval query can be executed at a time to prevent confusion caused by lag. The player is now notified the database is busy upon subsequent inspections.
Added error messages for when in inspection mode and trying to perform actions which would be cancelled to make it clear to the player.