Indexing
Added indexing on to the three main data managers for the main thing which inspection mode filters by for that object. For example, an index of posID has been created for block based interactions, and an entityID index has been created for entity based interactions. Indexes are created at the same time as tables. This is done synchronously and blocks the server thread until completion.
This adds a major performance improvement for retrieving entries with few interactions at the cost of storage space. This will be improved soon.
Updated database busy message.
Bumped version