Database Entry Merging
Database entries with identical data within a configurable timeframe are now merged with a count field. This only happens for data managers as the table managers should have unique entries anyway. This should act as a major space and efficiency optimisation against repeated actions such as cobblestone farms. Ideally we will want to reformat the count display in inspection mode to be localised and also refactor the combined block data system.