Initial 1.21.1 Port
It seems everything compiles and the game runs on both NeoForge and Fabric for 1.21.1. Some mixins almost certainly are broken and will probably crash when they are run and further testing is needed. Moved mixins around because of changes in 1.21.1 so some are common and others had to be split into specific modloader code.

On NeoForge we had to use a fork of the JDBC driver because it doesn't work with NeoForge yet, this is included in the repo files for now until a maven becomes available with it on.

A way around mixin crashes would be to set the require=0 flag on them to prevent a crash, this should also mean the default level mixin will catch them in the event of a failed mixin injection, which is a good fallback for mod compatibility also.

We probably won't continue support for 1.20.1 as this mod is still in development and not close to being complete and it's a lot of work to maintain the mixins here.